khadox_server/requirements.txt
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
mysqlclient==2.2.4
python-decouple==3.8
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
